backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
langchain==0.3.14
langchain-google-genai==2.0.8
langchain-chroma==0.2.0
chromadb==0.5.23
python-dotenv==1.0.1
pydantic==2.10.4
python-multipart==0.0.20
